pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monster-generator"
version = "0.1.0"
description = "Deterministic monster generator from sprite atlas"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0.0"]
license = {text = "MIT"}

[project.scripts]
monster-generator = "monster_generator.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
monster_generator = ["assets/*.png", "assets/*.xml"]